Move benchmark-related utility functions to benchmark_utils.py
The following benchmark-related functions were moved from
gcpy/util.py to gcpy/benchmark/modules/benchmark_utils.py:
- get_species_categories
- archive_species_categories
- get_lumped_species_definitions
- archive_lumped_species_definitions
- add_lumped_species_to_dataset

Also:

gcpy/benchmark/modules/benchmark_util.py
- Now import xarray
- Add constants for YAML files that are now contained in the
  gcpy/benchmark/modules folder.  These should not change very much,
  except when  benchmark categories or lumped species change.
- Updated Pydoc headers

gcpy/benchmark/modules/benchmark_funcs.py
- Now import several constants and functions from benchmark_utils.py
- Removed local constants for YAML files
- Updated function calls accordingly to be consistent with the new
  functions in benchmark_utils.py

CHANGELOG.md
- Updated accordingly
