+++ gcpy/benchmark/modules/benchmark_utils.py
@@ -4,12 +4,20 @@ TODO: Migrate other benchmark-specific utilities from gcpy/benchmark.py to here.
 """
 import os
 import numpy as np
+import xarray as xr
 import pandas as pd
 from gcpy import util
 from gcpy.constants import skip_these_vars
 
 # Suppress numpy divide by zero warnings to prevent output spam
 np.seterr(divide="ignore", invalid="ignore")
+
+# YAML files
+AOD_SPC = "aod_species.yml"
+BENCHMARK_CAT = "benchmark_categories.yml"
+EMISSION_SPC = "emission_species.yml"
+EMISSION_INV = "emission_inventories.yml"
+LUMPED_SPC = "lumped_species.yml"
 
 
 def make_output_dir(
@@ -75,9 +83,8 @@ def read_ref_and_dev(
     verbose    (bool) : Enable verbose output
 
     Returns:
-    --------
-    ref_data (xr.Dataset) : Data from the Ref model
-    dev_data (xr.Dataset) : Data from the Dev model
+    ref_data : xr.Dataset : Data from the Ref model
+    dev_data : xr.Dataset : Data from the Dev model
     """
     util.verify_variable_type(ref, (str, list))
     util.verify_variable_type(dev, (str, list))
@@ -110,16 +117,14 @@ def get_common_varnames(
     Returns an alphabetically-sorted list of common variables two
     xr.Dataset objects matching a given prefix.
 
-    Args:
-    -----
-    refdata (xr.Dataset ) : Data from the Ref model.
-    devdata (xr.Dataset ) : Data from the Dev model.
-    prefix  (str        ) : Variable prefix to match.
-    verbose (bool       ) : Toggle verbose printout on/off.
+    Args
+    refdata : xr.Dataset : Data from the Ref model.
+    devdata : xr.Dataset : Data from the Dev model.
+    prefix  : str        : Variable prefix to match.
+    verbose : bool       : Toggle verbose printout on/off.
 
-    Returns:
-    --------
-    varlist (list of str) : Sorted list of common variable names.
+    Returns
+    varlist : list       : Sorted list of common variable names.
     """
     vardict = util.compare_varnames(refdata, devdata, quiet=not verbose)
     varlist = [var for var in vardict["commonvars"] if prefix in var]
@@ -138,10 +143,11 @@ def print_sigdiffs(
     Appends a list of species showing significant differences in a
     benchmark plotting category to a file.
 
-    sigdiff_files (list|None) : List of files for signficant diffs output
-    sigdiff_list  (list     ) : List of significant differences to print
-    sigdiff_type  (str      ) : e.g. "sfc", "500hPa", "zm"
-    sigdiff_cat   (str      ) : e.g. "Oxidants", "Aerosols", "DryDep", etc.
+    Args
+    sigdiff_files : list|None : List of files for signficant diffs output
+    sigdiff_list  : list      : List of significant differences to print
+    sigdiff_type  : str       : e.g. "sfc", "500hPa", "zm"
+    sigdiff_cat   : str       : e.g. "Oxidants", "Aerosols", "DryDep", etc.
     """
     util.verify_variable_type(sigdiff_files, (list, type(None)))
     util.verify_variable_type(sigdiff_list, list)
@@ -163,9 +169,10 @@ def write_sigdiff(
     Appends a list of species showing significant differences in a
     benchmark plotting category to a file.
 
-    sigdiff_list (list ) : List of significant differences.
-    sigdiff_cat  (str  ) : e.g. "Oxidants", "Aerosols", "DryDep", etc.
-    sigdiff_file (str  ) : Filename to which the list will be appended
+    Args
+    sigdiff_list : list : List of significant differences.
+    sigdiff_cat  : str  : e.g. "Oxidants", "Aerosols", "DryDep", etc.
+    sigdiff_file : str  : Filename to which the list will be appended
     """
     util.verify_variable_type(sigdiff_list, list)
     util.verify_variable_type(sigdiff_cat, str)
@@ -188,16 +195,14 @@ def pdf_filename(
     """
     Creates the absolute path for a PDF file containing benchmark plots.
 
-    Args:
-    -----
-    dst        (str     ) : Root folder for benchmark output plots
-    collection (str     ) : e.g. "Aerosols", "DryDep", etc.
-    subdst     (str|None) : e.g. "AnnualMean", "Apr2019", etc.
-    plot_type  (str     ) : e.g. "Surface", "FullColumnZonalMean", etc.
+    Args
+    dst        : str      : Root folder for benchmark output plots
+    collection : str      : e.g. "Aerosols", "DryDep", etc.
+    subdst     : str|None : e.g. "AnnualMean", "Apr2019", etc.
+    plot_type  : str      : e.g. "Surface", "FullColumnZonalMean", etc.
 
-    Returns:
-    --------
-    pdf_path (str) : Absolute path for the PDF file containing plots
+    Returns
+    pdf_path   : str      : Absolute path for the PDF file containing plots
     """
     util.verify_variable_type(dst, str)
     util.verify_variable_type(collection, str)
@@ -216,9 +221,8 @@ def print_benchmark_info(
     """
     Prints which benchmark plots and tables will be generated.
 
-    Args:
-    -----
-    config (dict) : Inputs from the benchmark config YAML file.
+    Args
+    config : dict : Inputs from the benchmark config YAML file.
     """
     util.verify_variable_type(config, dict)
 
@@ -281,26 +285,13 @@ def get_geoschem_level_metadata(
     Reads a comma-separated variable (.csv) file with GEOS-Chem vertical
     level metadata and returns it in a pandas.DataFrame object.
 
-    Args:
-    -----
-    filename : str
-        Name of the comma-separated variable to read.
-        Default value: "__file__/GC_72_vertical_levels.csv"
+    Args
+    filename   : str          : Name of the comma-separated variable to read
+    search_key : str|None     : Returns metadata that matches this value
+    verbose    : bool         : Toggles verbose printout on or off
 
-    Keyword Args:
-    -------------
-    search_key : str
-        If present, will return metadata that matches this value.
-        Default: None
-
-    verbose : bool
-        Toggles verbose printout on (True) or off (False).
-        Default value: True
-
-    Returns:
-    --------
-    metadata : pandas.DataFrame
-        Metadata for each of the GEOS-Chem vertical levels.
+    Returns
+    metadata   : pd.DataFrame : Metadata for GEOS-Chem vertical levels
     """
     if filename is None:
         filename = os.path.join(
@@ -319,3 +310,186 @@ def get_geoschem_level_metadata(
     if search_key is None:
         return metadata
     return metadata[search_key]
+
+
+def get_lumped_species_definitions():
+    """
+    Returns lumped species definitions from a YAML file.
+
+    Returns
+    lumped_spc_dict : dict : Dictionary of lumped species
+    """
+    ifile = LUMPED_SPC
+    return util.read_config_file(
+        os.path.join(
+            os.path.dirname(__file__),
+            ifile,
+        ),
+        quiet=True
+    )
+
+
+def archive_lumped_species_definitions(
+        dst
+):
+    """
+    Archives lumped species definitions to a YAML file.
+
+    Args:
+    dst : str : Destination folder for YAML file output.
+    """
+    ofile = LUMPED_SPC
+    src = os.path.join(os.path.dirname(__file__), ofile)
+    copy = os.path.join(dst, ofile)
+    if not os.path.exists(copy):
+        print(f"\nArchiving {lumped_spc} in {dst}")
+        shutil.copyfile(src, copy)
+
+
+def add_lumped_species_to_dataset(
+        dset,
+        lspc_dict=None,
+        lspc_yaml="",
+        verbose=False,
+        overwrite=False,
+        prefix="SpeciesConcVV_",
+):
+    """
+    Function to calculate lumped species concentrations and add
+    them to an xarray Dataset. Lumped species definitions may be passed
+    as a dictionary or a path to a yaml file. If neither is passed then
+    the lumped species yaml file stored in gcpy is used. This file is
+    customized for use with benchmark simuation SpeciesConc diagnostic
+    collection output.
+
+    Args
+    dset      : xr.Dataset : Data prior to adding lumped species
+    lspc_dict : dict       : Species & scale factors for each lumped species
+    lspc_yaml : str        : YAML file w/ lumped species definitions
+    verbose   : bool       : Toggles verbose printout on or off.
+    overwrite : bool       : Overwrite existing species or raise an error
+    prefix    : str        : Prefix to prepend to lumped species names
+
+    Returns
+    dset      : xr.Dataset : Original species plus added lumped species
+    """
+
+    # Default is to add all benchmark lumped species.
+    # Can overwrite by passing a dictionary
+    # or a yaml file path containing one
+    assert not (
+        lspc_dict is not None and lspc_yaml != ""
+    ), "Cannot pass both lspc_dict and lspc_yaml. Choose one only."
+    if lspc_dict is None and lspc_yaml == "":
+        lspc_dict = get_lumped_species_definitions()
+    elif lspc_dict is None and lspc_yaml != "":
+        lspc_dict = util.read_config_file(lspc_yaml)
+
+    # Make sure attributes are transferred when copying dataset / dataarrays
+    with xr.set_options(keep_attrs=True):
+
+        # Get a dummy DataArray to use for initialization
+        dummy_darr = None
+        for var in dset.data_vars:
+            if prefix in var or prefix.replace("VV", "") in var:
+                dummy_darr = dset[var]
+                dummy_type = dummy_darr.dtype
+                dummy_shape = dummy_darr.shape
+                break
+        if dummy_darr is None:
+            msg = "Invalid prefix: " + prefix
+            raise ValueError(msg)
+
+        # Create a list with a copy of the dummy DataArray object
+        n_lumped_spc = len(lspc_dict)
+        lumped_spc = [None] * n_lumped_spc
+        for var, spcname in enumerate(lspc_dict):
+            lumped_spc[var] = dummy_darr.copy(deep=False)
+            lumped_spc[var].name = prefix + spcname
+            lumped_spc[var].values = np.full(dummy_shape, 0.0, dtype=dummy_type)
+
+        # Loop over lumped species list
+        for var, lspc in enumerate(lumped_spc):
+
+            # Search key for lspc_dict is lspc.name minus the prefix
+            cidx = lspc.name.find("_")
+            key = lspc.name[cidx+1:]
+
+            # Check if overlap with existing species
+            if lspc.name in dset.data_vars and overwrite:
+                dset.drop(lspc.name)
+            else:
+                assert(lspc.name not in dset.data_vars), \
+                    f"{lspc.name} already in dataset. To overwrite pass overwrite=True."
+
+            # Verbose prints
+            if verbose:
+                print(f"Creating {lspc.name}")
+
+            # Loop over and sum constituent species values
+            num_spc = 0
+            for _, spcname in enumerate(lspc_dict[key]):
+                varname = prefix + spcname
+                if varname not in dset.data_vars:
+                    if verbose:
+                        print(f"Warning: {varname} needed for {lspc_dict[key][spcname]} not in dataset")
+                    continue
+                if verbose:
+                    print(f" -> adding {varname} with scale {lspc_dict[key][spcname]}")
+                lspc.values += dset[varname].values * lspc_dict[key][spcname]
+                num_spc += 1
+
+            # Replace values with NaN if no species found in dataset
+            if num_spc == 0:
+                if verbose:
+                    print("No constituent species found! Setting to NaN.")
+                lspc.values = np.full(lspc.shape, np.nan)
+
+        # Insert the DataSet into the list of DataArrays
+        # so that we can only do the merge operation once
+        lumped_spc.insert(0, dset)
+        dset = xr.merge(lumped_spc)
+
+    return dset
+
+
+def get_species_categories(
+        benchmark_type="FullChemBenchmark"
+):
+    """
+    Returns the list of benchmark categories that each species
+    belongs to.  This determines which PDF files will contain the
+    plots for the various species.
+
+    Args
+    benchmark_type : str  : Specifies the type of the benchmark
+
+    Returns:
+    spc_cat_dict   : dict : Dictionary of categories and sub-categories
+    """
+    ifile = BENCHMARK_CAT
+    spc_cat_dict = util.read_config_file(
+        os.path.join(
+            os.path.dirname(__file__),
+            ifile,
+        )
+    )
+    return spc_cat_dict[benchmark_type]
+
+
+def archive_species_categories(
+        dst
+):
+    """
+    Writes the list of benchmark categories to a YAML file
+    for archival purposes.
+
+    Args:
+    dst  : str : Destination folder for YAML file output.
+    """
+    ofile = BENCHMARK_CAT
+    src = os.path.join(os.path.dirname(__file__), ofile)
+    copy = os.path.join(dst, ofile)
+    if not os.path.exists(copy):
+        print(f"\nArchiving {spc_categories} in {dst}")
+        shutil.copyfile(src, copy)
